# Improve Map minViewSure/maxViewSure