@treeowl 2024-08-06
The code seems clearer, which is good. We no longer need an explicit `NOINLINE`, which is good. I do not even slightly understand why this changes the Core in any way, which is bad. Do you think you could explain?