@treeowl 2024-08-06
I take that back. I get it now, I think. Since we've dropped the `NOINLINE`, GHC can expand it out one layer.