@meooow25 2024-08-06
Yes I think it makes sense. Previously the `go` was lifted out so we had a worker-wrapper for `go`. Then we had `minViewSure = goWrapper` with a `NOINLINE`.

```
-- RHS size: {terms: 1, types: 0, coercions: 0, joins: 0/0}
minViewSure [InlPrag=NOINLINE]
  :: forall k a. k -> a -> Map k a -> Map k a -> MinView k a
[GblId, Arity=4, Str=<1L><L><1L><1L>, Cpr=1, Unf=OtherCon []]
minViewSure = go10_rpA6
```

Now the separate `go` is gone. If we were worried about `minViewSure` inlining, the `NOINLINE` is unnecessary now because it is recursive and won't inline anyway.